Added a bit more readme and also added a summary message and stopwatch

+++ Program.cs
@@ -1,18 +1,26 @@
 ﻿// See https://aka.ms/new-console-template for more information
+using System.Diagnostics;
 using System.IO;
 
 Console.WriteLine("Hello, World!");
+var stopwatch = new Stopwatch();
+stopwatch.Start();
 
-int FOLDER_COUNT_MAX = 100;
-int FILE_COUNT_MAX = 100;
+const int FOLDER_COUNT_MAX = 100;
+const int FILE_COUNT_MAX = 100;
 string[] extensions = new[] { "XLSX", "DOCX", "TXT", "SQL", "PDF", "CS" };
+int totalFiles = 0;
+int totalFolders = 0;
 
 
 var path = Path.Combine(Directory.GetCurrentDirectory(), "files");
 
 #if DEBUG
 // auto cleanup for debugging!
-Directory.Delete(path, true);
+if (Directory.Exists(path))
+{
+    Directory.Delete(path, true);
+}
 #endif
 
 var random = new Random();
@@ -20,9 +28,12 @@ var random = new Random();
 for (var folderCount = 0; folderCount < FOLDER_COUNT_MAX; folderCount++)
 {
     var directory = Directory.CreateDirectory(GenerateFolderPath(path)).ToString();
-
+    totalFolders++;
     CreateFilesForFolder(directory);
 }
+
+stopwatch.Stop();
+Console.WriteLine($"Created a total of {totalFolders} folders with a combined {totalFiles} files in {stopwatch.Elapsed}");
 
 void CreateFilesForFolder(string folderPath, bool createNested = false)
 {
@@ -54,7 +65,6 @@ void GenerateFiles(string folderPath, int numberOfFiles)
     {
         var days = random.Next(1, 100);
         DateTime creationTime = DateTime.Now.AddDays(-random.Next(1, 100)).AddHours(random.Next(1, 100)).AddMinutes(random.Next(1, 100)); // randomise the file creation date anywhere from now to 100 days ago
-        Console.WriteLine($"Creating {numberOfFiles}");
         var extensionRandomiser = extensions[random.Next(extensions.Length)];
 
         var newFileName = $"{Guid.NewGuid()}.{extensionRandomiser}";
@@ -69,7 +79,8 @@ void GenerateFiles(string folderPath, int numberOfFiles)
 
         File.SetCreationTime(filePath, creationTime);
 
-        Console.WriteLine($"Create file {newFileName} in directory {folderPath} with a date of {creationTime}");
+        Console.WriteLine($"Created file {newFileName} in directory {folderPath} with a date of {creationTime}");
+        totalFiles++;
 
     }
 }
@@ -80,4 +91,8 @@ string GenerateFolderPath(string path)
     return newPath.ToString();
 }
 
-string CreateDirectory(string name) => Directory.CreateDirectory(Path.Combine(name, Guid.NewGuid().ToString()).ToString()).ToString();
+string CreateDirectory(string name)
+{
+    totalFolders++;
+    return Directory.CreateDirectory(Path.Combine(name, Guid.NewGuid().ToString()).ToString()).ToString();
+}
